.qsub files rewritten, .sh files removed, README.md added
The initial motivation for this commit was to rewrite `make_ucsc_references.qsub` so that it would work properly with all genomes.  Previously, individual .fa files would be retrieved from the chromosomes/ UCSC FTP directory, which exists for the human and mouse genomes, but not for others (e.g., the bovine genome).  This script has now been edited so that it retrieves a whole-genome .fa.gz file from the bigZips/ UCSC FTP directory (which should exist for all genomes) and splits it into individual .fa files.

In the process of this edit, both .qsub files were rewritten as follows:
* `getopt` is now used to process command-line arguments with more clarity and flexibility
* The job names are set within the qsub flags
* Any references to SCC-specific paths or environment variables have been removed
* Other edits were made throughout each script following ShellCheck review

The * .sh files were then removed because they were functionally redundant after the updates to the .qsub files.

README.md was also added to provide proper documentation.