Refactored make_ucsc_references.qsub and updated usage statements
This commit was initially prompted by the realization that the file 'md5sum.txt' does not necessarily contain a checksum for the file '[ucsc_build].fa.gz', but does appear to always contain one for the file '[ucsc_build].2bit'.  Furthermore, the 2bit file format is indexed, which makes it much easier to create FASTA files corresponding to sets of contigs.

Consequently, the code was refactored to use the .2bit file as input.  Also, to simplify things, the code at the end of the script was refactored into a loop that sequentially extracts each set of contigs (base, base_random, base_random_althap) to a FASTA file and creates derivative files from it.

Finally, the usage statements were slightly reformatted, and the unnecessary 'rm' commands to clean up the TMPDIR at the end of each qsub script were removed.