Fixed error in make_ucsc_refeerences.qsub
Following last commit, found the following error: was testing for presence of contig names in entire .chrom.sizes file (rather than the first column only) using regexes that are anchored to the start and end of the string.

While fixing this problem, I decided to refactor the code to store matching contigs in an array (as best practice would be to call 'grep' only once, not twice), and then also refactored the "base" code in the same manner for consistency.